v1
Day_2 - 
A floor has a center wooden section of size ‘n×m’, that has to be covered with carpet. The carpet has to be stitched using square pieces of cloth of length ‘a’.
Find the minimum number of cloths that must be stitched together to cover the wooden section. The carpet may cover an additional area outside the wooden floor. The input contains 3 integers, n m a, in the respective order.